requirements.txt
aiogram>=3.4.0
aiohttp-socks>=0.8.0
openai>=1.12.0
anthropic>=0.18.0
google-genai>=1.0.0
httpx[socks]>=0.27.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.3.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.2.0
tavily-python>=0.5.0
telegraph[aio]>=2.2.0
tensorflow>=2.15.0
tf-keras>=2.15.0
russtress>=0.1.3
num2words>=0.5.13
yt-dlp>=2024.0.0
youtube-transcript-api>=0.6.0
